remove container and statistics
